tests/wpt/web-platform-tests/trusted-types/TrustedTypePolicyFactory-createPolicy-nameTests.tentative.html

<!DOCTYPE html>
<script src="/resources/testharness.js" ></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/helper.sub.js"></script>
<body>
<script>
  //Policy name test
  test(t => {
    let policy = window.TrustedTypes.createPolicy('SomeName', { createHTML: s => s } );
    assert_true(policy instanceof TrustedTypePolicy);
    assert_equals(policy.name, 'SomeName');
  }, "policy.name = name");

  //Duplicate names test
  test(t => {
    assert_throws(new TypeError(), _ => {
     window.TrustedTypes.createPolicy('SomeName', { createURL: s => s } );
    });
  }, "duplicate policy name attempt throws");

  //Retrieve policy names tests
  test(t => {
    let policy = window.TrustedTypes.createPolicy('SomeOtherName', { createURL: s => s } );
    let names = window.TrustedTypes.getPolicyNames();
    assert_true(names.includes('SomeName'));
    assert_true(names.includes('SomeOtherName'));
  }, "Retrieving policy names");
</script>
